requirements.txt
amqp==2.4.2
asgiref==3.1.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.5.1
Automat==0.7.0
Babel==2.5.3
backports.csv==1.0.7
baidu-aip==2.2.13.0
beautifulsoup4==4.7.1
billiard==3.6.0.0
celery==4.3.0
certifi==2018.4.16
channels==2.2.0
chardet==3.0.4
colorama==0.3.9
constantly==15.1.0
daphne==2.3.0
decorator==4.4.0
defusedxml==0.6.0
Delorean==1.0.0
dj-dajaxice==1.0
Django==2.0.5
django-bootstrap-pagination==1.6.4
django-pagination==1.0.7
django-redis==4.10.0
django-stdimage==3.1.1
djangorestframework==3.9.0
docopt==0.6.2
docutils==0.14
dukpy==0.2.2
dwebsocket==0.5.10
et-xmlfile==1.0.1
fire==0.1.3
future==0.17.1
geohash2==1.1
google==2.0.2
humanize==0.5.1
hyperlink==19.0.0
idna==2.7
imageio==2.5.0
incremental==17.5.0
javascripthon==0.10
jdcal==1.4.1
Jinja2==2.10
jupyter-echarts-pypkg==0.1.2
kombu==4.5.0
lml==0.0.2
logzero==1.5.0
lxml==4.3.3
macropy3==1.1.0b2
MarkupSafe==1.1.0
MyQR==2.3.1
mysqlclient==1.3.12
numpy==1.15.4
odfpy==1.4.0
openpyxl==2.6.2
paho-mqtt==1.4.0
pandas==0.23.4
Pillow==5.1.0
progress==1.5
progressbar2==3.37.1
protobuf==3.8.0
psycopg2==2.7.7
py==1.8.0
PyCRC==1.21
pycurl==7.43.0.2
pyecharts==0.5.11
pyecharts-javascripthon==0.0.6
pyecharts-jupyter-installer==0.0.3
pyecharts-snapshot==0.1.8
Pygments==2.2.0
PyHamcrest==1.9.0
pytesseract==0.2.6
python-dateutil==2.7.2
python-memcached==1.59
python-utils==2.3.0
pytz==2018.4
PyYAML==5.1
qrcode==6.0
redis==3.2.1
requests==2.19.1
retry==0.9.2
scikit-learn==0.20.3
scipy==1.2.1
selenium==3.141.0
simplejson==3.16.0
six==1.11.0
sklearn==0.0
soupsieve==1.9.1
tablib==0.13.0
tinify==1.5.1
Twisted==19.2.0
txaio==18.8.1
tzlocal==1.5.1
uiautomator2==0.2.2
urllib3==1.23
vine==1.3.0
whichcraft==0.5.2
win32compat==221.26
win32core==221.36
xlrd==1.1.0
xlwt==1.3.0
xmlrpc2==0.3.1
zope.interface==4.6.0
